requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2==2.8.6
pytz==2019.3
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.0.4
certifi==2019.11.28
idna==2.8
chardet==3.0.4
requests==2.22.0
urllib3==1.25.8


